Plugin tab: Reduce the database server's strain a wee bit more by applying a thoroughly researched (but, sorry to say, irreversible and lossy) compression algorithm to the data constituting the plugin list.
As this is probably my last commit comment for 2007, I'd like to make use of this opportunity to wish a Happy New Year to everybody living on a Gregorian schedule on behalf of the group and ourselves. "On behalf of the group and ourselves" is, as an aside, a phrase which was coined by the late John Lennon during the live recording of "Get Back" on the Apple building's roof top which was later on released as a track on the last Beatles album "Let It Be". Paul McCartney complemented their salute by moaning "Thanks, Mo", directed at Maureen Cox Starkey, who was Ringo's wife from 1965 to 1975. But I digress...

git-svn-id: http://textpattern.googlecode.com/svn/development/4.0@2757 2fea6a4d-a838-0410-917b-93a53c48e9d2