Introduces parse_page() and fetch_page().
Allows fetching and parsing page templates similarly to
forms. These functions do all the needed processing on the
background.

Now Textpattern also allows page templates that only
contain a single empty byte such as '0' or an empty string,
instead of displaying unexplained 404 error page. This error
page is displayed when the template is missing.

The fetching method used for pages can be extended with
a 'page.fetch' callback event, similarly to forms.

Closes issue #363

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@5228 2fea6a4d-a838-0410-917b-93a53c48e9d2