Fix parent title searches that start with numbers.
In short, that doesn't do what it think it does. Any numeric
like value gets evaluated as TRUE. Particularly, in PHP, any
string that starts with a number becomes a non-zero
integer when casting, unless the value at the start is
considered as 'zero'.

Thus, if this check really needs to be there, it requires
casting and also comparing it to the original value in string
representation. Identical to avoid type juggling that
otherwise occurs again.

git-svn-id: http://textpattern.googlecode.com/svn/development/4.x@5254 2fea6a4d-a838-0410-917b-93a53c48e9d2