Unlimited custom fields support objects
Includes a new DB library function safe_updin() which behaves a little like safe_upsert() in that it performs an update and then an insert if necessary, but it works with multicolumn WHERE clauses. Pass in arrays of tuples to set / test and they will be merged correctly depending on the action reuqired.

Also includes a new global constant $txpnow which holds a single value for "right now". This avoids the potential 'wandering time' issue, which could appear due to the time taken to process and render a page. It anchors all meta inserts/updates to a single, unified timestamp.