Abort update scripts on errors.
I've added an "updateErrorHandler" which basically does nothing but:

1. temporarily set production status to 'debug' (needed, because if the update fails, you can't get into preferences and you need the backtrace to diagnose the error).
2. call the adminErrorHandler (saves duplicating code)
3. throw an exception (to get out of the loop of running update scripts)

The 'version' pref is set to the last database update script that ran succesfully, so it can start again on the next attempt at the file where the error occurred.

There's an array of update versions at the top of _update which contains all the numbers of the update files. Note that $dbversion is the TXP version where the database structure last changed, which can be equal to the installed TXP version, but can also be older if a newer TXP version didn't introduce database changes. No need to create an empty _to_4.x.x update file if nothing changes in the database.