Improve comment error handling
Put all the cleanup of comment form field values into the getComment() function. The advantage being that we get the same type of cleanup everwhere we use these values.

Evaluate required fields during preview at a much earlier stage, allowing errors to be shown even if the errors are to be displayed before the comment form fields. Fixes issue #623.

Add language string: comment_duplicate, which is an error message indicating that the comment was rejected, because the message was a duplicate (already exists in database).