Shield plugins from internal prerequisites
The upload_form() function has an optional class parameter which is
appended to a preset 'upload-form' class name. All core functions employ a
space before passing in this value so the additional class name is
appended without issue. Plugin authors should not be burdened with this
trivia so the space is not assumed and used automatically when required.
Thanks, jakob.