Suppress errors if old files cannot be deleted
Inelegant, but if permissions are cranked down low or the upgrade is run more than once for whatever reason (beta test, failure, etc) then it's better to silently absorb the errors than to unceremoniously halt. It's no biggie if the files stick around.

Longer term, it'd be better to detect this and offer some advice that deletion couldn't complete, but in the absence of a Flash Message notification system on first login, this'll have to do.