Introduce admin side language per-user
Fixes #644.

Note this opens up the Languages panel to everybody so they can choose their own language. A new priv has thus been created (lang.edit) that permits privileged users to be able to manage the front-of-house language and to install/remove entire language packs. Textpack upload is currenty available to all. We can review this if necessary.

A global, user-overridable constant to define the default language has also been added. This does NOT affect the setup routine at present, because only en-gb is shipped. When the entire language set is distributed in core, the default language constant can be rolled out to setup too.

This is the first part of the language panel revamp, so the UI for non-privileged users (when Show details is on) is bare. When the overall panel layout is improved, this will be addressed.