Plugin help rendered LTR unless otherwise stated
Closes #590. Thanks, phiw13.

Plugin help block can optionally now start with a 'comment line' that
begins "#@language" followed by the language-country identifier. e.g.
"#@language he-il" (without the quotes) would indicate the help text is
written in Hebrew and, as such, will be rendered RTL irrespective of user
interface language, as long as the nominated language is installed.

The identifier MUST occur on its own line, and be the first line of the
help text. It will be automatically stripped from the output when viewing
plugin help.

In the absence of a language marker, English (en-gb) is assumed.