Render pophelp links for Textfilters
See issue #887.

Note that this commit has the following ramifications:

* Side help twisty is now gone from the Write panel.
* article > sidehelp callback has been removed.
* Textfilter HTML help rendering system has been removed in favour of a simple URL to a (possibly third party) resource to supply help content. Such content is rendered inside our popHelp container, currently.
* Anyone supplying their own URL to popHelp() will have that content rendered inside our popups, which makes it easier to use without requiring the admin_help callback for simple URL content. Very handy for custom prefs.
* i18n of help content must be handled by the third party site. To assist with this, we could (but don't currently) pass the current in-use language designator to the third party URL (perhaps in headers?) For Textile this is currently unnecessary, since our chosen help URL endpoint (txstyle.org) is in English only.

As a result, the current implementation of Markdown by wet (see https://github.com/rwetzlmayr/wet_textfilter_markdown) will need its getHelp() function updating to return a single URL.

If we can find a sensible way of detecting if the content passed to popHelp is either a) a URL, b) a Txp help topic item string (e.g. `section_default`), or c) raw markup, we could reinstate the ability for authors to supply their own HTML content.

This could be done easily now for Textfilter content, but it would be good to be able to do this for all popHelp content, as it offers more flexibility. A possible solution might be to flag the content type to pophelp as a separate argument. e.g. it defaults to 'item' (to be read from RPC/local help topic) but could be set to 'url' to render a third party resource, or 'html' for raw HTML content, or 'json', or ...

Investigation ongoing.