Introduce theme description column
Used for supplying theme release information and supplementary actions, such as: set certain prefs, install plugins, caveats, typical setups, and so forth.

There is no way to display the information right now (besides the theme's Edit step) as it doesn't make sense to put it in the listtable. Perhaps a modal popup is suitable when clicking some other field (e.g. Title?), or adding an (i)nfo link somewhere in each row.

Note the column is varchar(16383) - as TEXT fields aren't cacheable - which gives 16KB of space for unibyte characters and 4KB of full multi-byte space. This increases the minimum system requirement for MySQL to v5.0.3.