More correct privs handling on Images panel
Prior to this commit, any unauthorized activity was met with a green status message saying "Restricted area". Now a proper privs violation is thrown and the page doesn't render.

Also, if trying to edit a non-existent image, just the footer was rendered. Now, an 'Unknown image' error is thrown as a status message and the user is dumped back to the image list.