Update lastmod when pasting Textpacks too
The fixed 'date in the past' was a legacy feature to prevent the RPC server from getting confused. If you update strings now, they get the stamp of the time they were updated.

This should not interfere with the UI widgets on the Languages panel as the file system will still be "older" than the pasted strings. To be able to overwrite strings from file via the UI, you need to touch or update the file in the lang directory as usual.