Introduce setOutput() to control tag rendering
From the point this method is called, any calls to setBool() and any calls to render() will take into account the defined output scheme - either html5 (the default) or xhtml.

This method acts as a global toggle for all tags and can be switched at any time to alter the output.

setAtts() is unaffected and will always render long name="value" attributes. If the attribute is set to 'true' the name of the attribute will be used as its value.