Workaround array conversion errors
At the expense of keeping NULL values in the database when the corresponding values are 'cleared'. Might be a better way to handle this to avoid:

a) SQL errors when storing empty '' values in datetime columns.
b) PHP errors when $thisContent is an array but the render() case expects a string.