Added some security HTTP headers
These values can be applied by default to all textpattern installation. Unless there is a specific plugin that explicitly needs these security settings to be relaxed/tweaked these can help improve the security of a default installation. 'unsafe-inline' in csp of script-src is a reminder that it can be eventually fixed to improve security. I use these and more tweaked settings with reporting on my website. The CSP settings can be verified with csp-evaluator.withgoogle.com.