Introduce pref family column
A sub-genre of event, this officially allows a more granular way of splitting prefs up by event. Separating the event name by '.' continues to work for plugins, but event->family doesn't require individual privileges for each family.

When using set_pref() and its ilk, you can pass an array to the $event param containing the array(event,family) in which the pref should reside. When displayed on the Prefs panel, the family will be used as a heading (h3) with a class of the family name, and all prefs beneath it also have the family's name as their class.