Not much use now but... why not callback
Permits plugins to have a say so they can partially alter/augment the supprted image types (minus thumbnails as that's governed by the hard-coded check_gd() function). This change does permit a plugin to revoke .swf for good, which could be handy for those that want to expressly forbid it.