Remove GRANTS checking during setup for now
Seems it's way more involved than expected, as different systems/admins/flavours of MySQL can store data in the INFORMATION_SCHEMA.USER_PRIVILEGES table in different ways.

For expedience, axe this feature for now but reserve the right to reinstate it in future if we find a universal way (perhaps in PDO) to determine the user rights are sufficient.