Implemented crop function for SVG thumbnails
The cropping function of SVG images largely simulates the behavior of other image types, since the same function is used to determine the image size of the thumbnails based on user input. However, if both width and height are specified and the ratio between them does not match the ratio of the original image, the thumbnail is not distorted but centered in the viewbox while maintaining the original image ratio.