Introduce field size constraints via JavaScript
Add a data-max-chars attributes to any VARCHAR input field and the counter will be invoked. It tries as best as it can to respect UTF-8 character byte counts. YMMV with emoji! Improvements to the JavaScript counter welcomed.

Any typed/pasted input that would violate the character count will be unceremoniously truncated, without regard to word boundaries. Maybe a future enhancement would be to add a 'Remaining characters' count somewhere near the input.

See #1800 and #1801.