Add name maxlength constraint for Styles
Interestingly, the skin sanitizer limits the stored name to 63 chars, while the database storage can theoretically hold up to 255 chars (the maxlength of the input control). Anything above the 63rd character will be truncated without warning.

Whether this was done for INDEX purposes or if it is a requirement of some operating systems that filenames be significantly shorter than 255 chars isn't clear.

See #1800.