Create .htaccess
Commented `.htaccess` to limit 'raw' image access if there's no default document.

Aside: toolbelt removes instances of `.gitignore` as part of packaging, so there's an empty directory when an archive is expanded. If the expanded archive contents are added to a VCS that ignores empty directories (hola, Git), it won't track the empty directory, and that subsequently throws an error in Diagnostics where the default `images` directory is missing. This is addressed in a somewhat clunky way with this file.