More robust database layer
Trap errors using try... catch around safe_query. Although not recommended, this at least throws fewer fatal errors when things go sideways.

Additionally, attempt to limit the size of textareas in the Styles and Pages panels. This is an approximation at best because the theoretical limit of TEXT and MEDIUMTEXT always exceeds the practical limit due to multi-byte characters, storage overheads, sizes of other VARCHAR values in the row, and so forth. But at least the try... catch returns a proper error message via announce(), even if it's not particularly informative.

More problematic are MEDIUMTEXT fields. In MySQL, these can hold a maximum of 16MB but there are other limits that will be hit before the storage is maxed out. One is `max_allowed_packet`. A new constant has been added to permit this to be modified, and it defaults to 100MB (because the plugins table has 5x MEDIUMTEXT fields = 80MB theoretical maximum, plus sundry column data).

The other limit is PHP's `post_max_size` which cannot be altered in a runtime PHP environment (only by editing php.ini or modifying .htaccess). This defaults to 8MB and is usually hit way sooner than any individual column storage limit. It's difficult to defend against, but the Ajax response will throw a messy warning:

POST Content-Length of nnnn bytes exceeds the limit of 8388608 bytes in Unknown on line 0

and will not return an announce() message.

See #2005.