Introduce Theme prefs + Textpacks
Add prefs.json to your admin-theme folder, with a valid JSON structure similar to core.prefs (there's no need for 'event' or 'type' attributes because they're assumed to be the theme name and THEME_PREF type). Use "private": "1" for a per-user pref, or omit it for global prefs.

Add textpack.txp with valid strings that match your prefs to label them. Supports 'instructions_' too. The 'owner' is assumed to be the theme name so you don't need to specify it in your pack.

All strings and prefs are "namespaced" in the database, prefixed with: "your_theme_name|"

So, for example, a hive neutral pref might be hiveneutral|grid_columns and you can use:

get_pref($this->name.PREF_THEME_DELIMETER.'grid_columns')

to fetch it. If this system works well, a Theme::get_pref('grid_columns') method would make things simpler and handle the delimiter and prefix behind the scenes.

NOTE: the previous behaviour of prefs.json (for injecting custom public JavaScript prefs and style blocks) has been renamed custom.json instead.